Finances projet : heures prévues — ne plus tronquer les durées de phase
duration_days est décimal : la planification autorise les demi-journées
(0.5 j = 4 h) et fractions (1.25 j = 10 h). Le calcul des heures prévues
faisait int(duration_days), ce qui ignorait toute phase < 1 j et
arrondissait 1.75 j à 1 j. On utilise désormais la valeur décimale —
les heures prévues reflètent la planification réelle journée par journée.

+++ backend/app/api/v1/endpoints/project_finances.py
@@ -314,7 +314,13 @@ async def _compute_finances(
                 )
 
         for ph in phases:
-            days = int(ph.duration_days or 0)
+            # duration_days est décimal (Numeric 6,2) : la planification
+            # autorise les demi-journées et fractions — 0.5 j = 4 h,
+            # 1.25 j = 10 h. On NE tronque PAS : un int() ferait
+            # disparaître les demi-journées (0.5 → 0, phase ignorée) et
+            # arrondirait 1.75 j à 1 j. Les heures prévues doivent
+            # refléter la planification réelle, journée par journée.
+            days = float(ph.duration_days or 0)
             if days <= 0:
                 continue
             hours = days * 8
